Fix Temperature program: change input method from nextInt() to nextDouble() for accurate temperature reading Add Time program: implement functionality to convert seconds into hours, minutes, and seconds

+++ JavaBasics/Level-1 WarmUp/Temperature.java
@@ -5,7 +5,7 @@ public class Temperature {
     public static void main(String [] args){
     Scanner sc = new Scanner(System.in);
     System.out.println("Enter temperature in Celsius: ");
-    double C = sc.nextInt();
+    double C = sc.nextDouble();
 
     double F = (C * 9/5) + 32;
     System.out.println("The temperature is: " + F + "°F");
